Added old age, better configuaration for creature types, dynamic texture loading.

+++ SoupV2/Simulation/Statistics/AbstractSimulationInfo.cs
@@ -6,7 +6,7 @@ namespace SoupV2.Simulation.EventsAndInfo
 {
     public abstract class AbstractSimulationInfo
     {
-        public uint TimeStamp { get; set; }
+        public float TimeStamp { get; set; }
         
     }
 }
